Move lookup classes to subpackage.
