+++ main/src/main/java/com/github/topisenpai/lavasrc/mirror/MirroringAudioSourceManager.java
@@ -1,5 +1,6 @@
 package com.github.topisenpai.lavasrc.mirror;
 
+import com.github.topisenpai.lavasrc.mirror.lookup.MirroringAudioTrackLookup;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
 import com.sedmelluq.discord.lavaplayer.source.AudioSourceManager;
 import com.sedmelluq.discord.lavaplayer.tools.DataFormatTools;
