+++ main/src/main/java/com/github/topisenpai/lavasrc/mirror/lookup/MirroringAudioTrackLookup.java
@@ -1,5 +1,6 @@
-package com.github.topisenpai.lavasrc.mirror;
+package com.github.topisenpai.lavasrc.mirror.lookup;
 
+import com.github.topisenpai.lavasrc.mirror.MirroringAudioTrack;
 import com.sedmelluq.discord.lavaplayer.track.AudioItem;
 
 import java.util.function.Function;
